backend/requirements.txt
django
Django>=5.2,<5.3
psycopg2-binary>=2.9.9
djangorestframework
django-cors-headers